Move Dynatrace metrics exporter for API v1 to a separate package
